+++ packages/terra-demographics-banner/src/DemographicsBanner.jsx
@@ -5,31 +5,50 @@ import DemographicsBannerDisplay from './DemographicsBannerDisplay';
 
 const identifierShape = {
   /**
-  * The label of the identifier: i.e.: `SSN`
-  */
+   * The label of the identifier: i.e.: `SSN`
+   */
   label: PropTypes.string,
   /**
-  * The value to display for the identifier. i.e.: `111-22-3333`
-  */
+   * The value to display for the identifier. i.e.: `111-22-3333`
+   */
   value: PropTypes.string,
   /**
-  * The aria-label for the identifier. i.e.: `Social Security Number 1 1 1 2 2 3 3 3 3`
-  */
+   * The aria-label for the identifier. i.e.: `Social Security Number 1 1 1 2 2 3 3 3 3`
+   */
   ariaLabel: PropTypes.string,
 };
 
 const stringShape = {
   /**
-  * The value to display. i.e.: `03/15/2020`
-  */
+   * The value to display. i.e.: `03/15/2020`
+   */
   value: PropTypes.string,
   /**
-  * The aria-label for the value. i.e.: `March 15th, 2020`
-  */
+   * The aria-label for the value. i.e.: `March 15th, 2020`
+   */
   ariaLabel: PropTypes.string,
 };
 
+const administrativeSexShape = {
+  /**
+   * The sex to display. i.e.: `M`
+   */
+  value: PropTypes.string,
+  /**
+   * The aria-label for the sex. i.e.: `Male`
+   */
+  ariaLabel: PropTypes.string,
+  /**
+   * Whether or not the administrative sex is different from the birth sex. If different, an icon will be displayed to denote the mismatch in sex.
+   */
+  isBirthSexMismatch: PropTypes.bool,
+};
+
 const propTypes = {
+  /**
+   * Administrative sex of the Person. This will be displayed on the demographics banner.
+   */
+  administrativeSex: PropTypes.shape(administrativeSexShape),
   /**
    * Age of the person.
    */
@@ -46,14 +65,6 @@ const propTypes = {
    * The person's deceased date. Will display the banner as deceased if this value is provided.
    */
   deceasedDate: PropTypes.shape(stringShape),
-  /**
-   * Gender of the Person. This will be displayed on the demographics banner.
-   */
-  gender: PropTypes.string,
-  /**
-   * The gender label used for accessibility. Unlike the `Gender` prop, this label will not be displayed. It will be read in screen readers if provided.
-   */
-  genderAria: PropTypes.string,
   /**
    * The person's gestational age.
    */
@@ -100,11 +111,11 @@ const propTypes = {
 };
 
 const defaultProps = {
+  administrativeSex: {},
   age: undefined,
   applicationContent: null,
   dateOfBirth: {},
   deceasedDate: {},
-  gender: undefined,
   gestationalAge: null,
   identifiers: undefined,
   isActive: false,
@@ -117,21 +128,20 @@ const defaultProps = {
 };
 
 const DemographicsBanner = ({
+  administrativeSex,
   age,
   dateOfBirth,
-  gender,
-  genderAria,
   intl,
   personName,
   ...customProps
 }) => (
   <DemographicsBannerDisplay
     {...customProps}
+    administrativeSex={administrativeSex}
     age={age}
     dateOfBirth={dateOfBirth}
-    gender={gender}
-    genderAria={genderAria}
     personName={personName}
+    administrativeSexFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.administrativeSexFullText' })}
     dateOfBirthLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.dateOfBirth' })}
     dateOfBirthFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.dateOfBirthFullText' })}
     deceasedDateLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.deceased' })}
@@ -139,6 +149,7 @@ const DemographicsBanner = ({
     gestationalAgeFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.gestationalAgeFullText' })}
     postMenstrualAgeLabel={intl.formatMessage({ id: 'Terra.demographicsBanner.postMenstrualAge' })}
     postMenstrualAgeFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.postMenstrualAgeFullText' })}
+    sexMismatchFullText={intl.formatMessage({ id: 'Terra.demographicsBanner.sexNotIdentical' })}
   />
 );