+++ packages/terra-demographics-banner/src/DemographicsBannerDisplay.jsx
@@ -8,31 +8,54 @@ import LargeDemographicsBannerDisplay from './_LargeDemographicsBannerDisplay';
 
 const identifierShape = {
   /**
-  * The label of the identifier: i.e.: `SSN`
-  */
+   * The label of the identifier: i.e.: `SSN`
+   */
   label: PropTypes.string,
   /**
-  * The value to display for the identifier. i.e.: `111-22-3333`
-  */
+   * The value to display for the identifier. i.e.: `111-22-3333`
+   */
   value: PropTypes.string,
   /**
-  * The aria-label for the identifier. i.e.: `Social Security Number 1 1 1 2 2 3 3 3 3`
-  */
+   * The aria-label for the identifier. i.e.: `Social Security Number 1 1 1 2 2 3 3 3 3`
+   */
   ariaLabel: PropTypes.string,
 };
 
 const stringShape = {
   /**
-  * The value to display. i.e.: `03/15/2020`
-  */
+   * The value to display. i.e.: `03/15/2020`
+   */
   value: PropTypes.string,
   /**
-  * The aria-label for the value. i.e.: `March 15th, 2020`
-  */
+   * The aria-label for the value. i.e.: `March 15th, 2020`
+   */
   ariaLabel: PropTypes.string,
 };
 
+const administrativeSexShape = {
+  /**
+   * The sex to display. i.e.: `M`
+   */
+  value: PropTypes.string,
+  /**
+   * The aria-label for the sex. i.e.: `Male`
+   */
+  ariaLabel: PropTypes.string,
+  /**
+   * Whether or not the administrative sex is different from the birth sex. If different, an icon will be displayed to denote the mismatch in sex.
+   */
+  isBirthSexMismatch: PropTypes.bool,
+};
+
 const propTypes = {
+  /**
+   * Administrative sex of the Person. This will be displayed on the demographics banner.
+   */
+  administrativeSex: PropTypes.shape(administrativeSexShape),
+  /**
+   * Text description of the Administrative Sex label that is read by screen readers
+   */
+  administrativeSexFullText: PropTypes.string,
   /**
    * Age of the person.
    */
@@ -61,14 +84,6 @@ const propTypes = {
    * Label to display for the deceased date
    */
   deceasedDateLabel: PropTypes.string,
-  /**
-   * Gender of the Person. This will be displayed on the demographics banner.
-   */
-  gender: PropTypes.string,
-  /**
-   * The gender label used for accessibility. Unlike the `Gender` prop, this label will not be displayed. It will be read in screen readers if provided.
-   */
-  genderAria: PropTypes.string,
   /**
    * The persons gestational age.
    */
@@ -121,16 +136,22 @@ const propTypes = {
    */
   postMenstrualAgeLabel: PropTypes.string,
   /**
-  * Text Description of the Post Menstrural Age label that is read by screen readers
+   * Text description of the Post Menstrural Age label that is read by screen readers
    */
   postMenstrualAgeFullText: PropTypes.string,
   /**
    * The persons preferred first name if they have one.
    */
   preferredFirstName: PropTypes.string,
+  /**
+   * Text description that is read by screen readers when the Administrative Sex and Birth Sex are not identical.
+   */
+  sexMismatchFullText: PropTypes.string,
 };
 
 const defaultProps = {
+  administrativeSex: {},
+  administrativeSexFullText: 'Administrative Sex',
   age: '',
   applicationContent: null,
   dateOfBirth: {},
@@ -138,7 +159,6 @@ const defaultProps = {
   dateOfBirthFullText: 'Date of Birth',
   deceasedDate: {},
   deceasedDateLabel: 'Deceased',
-  gender: '',
   gestationalAge: null,
   gestationalAgeLabel: 'GA',
   gestationalAgeFullText: 'Gestational Age',
@@ -151,6 +171,7 @@ const defaultProps = {
   postMenstrualAgeLabel: 'PMA',
   postMenstrualAgeFullText: 'Post Menstrual Age',
   preferredFirstName: null,
+  sexMismatchFullText: 'The documented sex values are not identical.',
 };
 
 const DemographicsBannerDisplay = props => (