+++ packages/terra-demographics-banner/src/terra-dev-site/doc/example/DemographicsBannerConfidential.jsx
@@ -8,7 +8,7 @@ const DemographicsBannerConfidential = () => (
     applicationContent={<span>St. Johns Kansas City West Wing Room Cardiac Arrest Unit 253</span>}
     age="25 Years"
     dateOfBirth={{ value: 'May 9, 1993', ariaLabel: 'May 9, 1993' }}
-    gender="Male"
+    administrativeSex={{ value: 'Male', ariaLabel: 'Male', isBirthSexMismatch: false }}
     identifiers={[
       {
         label: 'MRN',