+++ packages/terra-demographics-banner/src/terra-dev-site/doc/selectable-demographics-banner/SelectableDemographicsBannerAdditionalAges.jsx
@@ -8,7 +8,7 @@ const SelectableDemographicsBannerAdditionalAges = () => (
     applicationContent={<span>St. Johns Kansas City West Wing Room Cardiac Arrest Unit 253</span>}
     age="10 Weeks"
     dateOfBirth={{ value: 'October 25, 2018', ariaLabel: 'October 25, 2018' }}
-    gender="Male"
+    administrativeSex={{ value: 'Male', ariaLabel: 'Male', isBirthSexMismatch: false }}
     gestationalAge="33 Weeks"
     identifiers={[
       {