+++ packages/terra-demographics-banner/src/terra-dev-site/test/demographics-banner/selectable-demographics-banner/DeceasedSelectableDemographicsBanner.test.jsx
@@ -18,8 +18,7 @@ export default () => {
         age="27 Years"
         dateOfBirth={{ value: 'May 9, 1993', ariaLabel: 'May 9, 1993' }}
         deceasedDate={{ value: 'March 12, 2020', ariaLabel: 'March 12, 2020' }}
-        gender="M"
-        genderAria="Male"
+        administrativeSex={{ value: 'M', ariaLabel: 'Male', isBirthSexMismatch: false }}
         identifiers={[
           {
             label: 'MRN',